Move final remainders into stages
