+++ ostrich/ostrich.py
@@ -33,254 +33,6 @@ import utils
 ARGS = None
 
 
-def expand_ironic_netblock(r):
-    net = ipaddress.ip_network(r.complete['ironic-ip-block'])
-    hosts = []
-    for h in net.hosts():
-        hosts.append(str(h))
-
-    return net, hosts
-
-
-def stage7_user_variables(r, **kwargs):
-    """Configure user variables with all our special things."""
-
-    nextsteps = []
-
-    if r.complete['http-proxy'] and r.complete['http-proxy'] != 'none':
-        # This is the more permanent way of doing this
-        local_servers = 'localhost,127.0.0.1'
-        if r.complete['local-cache'] != 'none':
-            local_servers += ',%s' % r.complete['local-cache']
-
-        if r.complete['osa-branch'] == 'stable/mitaka':
-            nextsteps.append(
-                steps.FileAppendStep(
-                    'proxy-environment-via-ansible',
-                    '/etc/openstack_deploy/user_variables.yml',
-                    (('\n\n'
-                      'no_proxy_env: "%(local)s,{{ '
-                      'internal_lb_vip_address }},{{ external_lb_vip_address }},'
-                      '{%% for host in groups[\'all_containers\'] %%}'
-                      '{{ hostvars[host][\'container_address\'] }}'
-                      '{%% if not loop.last %%},{%% endif %%}{%% endfor %%}"\n'
-                      'global_environment_variables:\n'
-                      '  HTTPS_PROXY: "%(proxy)s"\n'
-                      '  https_proxy: "%(proxy)s"\n'
-                      '  HTTP_PROXY: "%(proxy)s"\n'
-                      '  http_proxy: "%(proxy)s"\n'
-                      '  NO_PROXY: "{{ no_proxy_env }}"\n'
-                      '  no_proxy: "{{ no_proxy_env }}"')
-                     % {'proxy': r.complete['http-proxy'],
-                        'local': local_servers}),
-                    **kwargs)
-                )
-
-    nextsteps.append(
-        steps.FileAppendStep(
-            'osa-debug-mode',
-            '/etc/openstack_deploy/user_variables.yml',
-            '\n\ndebug: true\nverbose: true',
-            **kwargs)
-        )
-
-    nextsteps.append(
-        steps.FileCreateStep(
-            'lxc-hosts-apt-keep-configs',
-            '/etc/ansible/roles/lxc_hosts/templates/apt-keep-configs.j2',
-            """Dpkg::Options {
-   "--force-confdef";
-   "--force-confold";
-}""",
-            **kwargs)
-        )
-
-    nextsteps.append(
-        steps.SimpleCommandStep(
-            'lxc-hosts-apt-keep-configs-enable',
-            """sed -i -e '/- name: Update container resolvers/ i \\- name: Always keep modified config files\\n  template:\\n    src: apt-keep-configs.j2\\n    dest: "{{ lxc_container_cache_path }}/{{ item.chroot_path }}/etc/apt/apt.conf.d/00apt-keep-configs"\\n  with_items: lxc_container_caches\\n  tags:\\n    - lxc-cache\\n    - lxc-cache-update\\n\\n'  /etc/ansible/roles/lxc_hosts/tasks/lxc_cache_preparation.yml""",
-            **kwargs)
-        )
-
-    if r.complete['osa-branch'] == 'stable/mitaka':
-        nextsteps.append(steps.PatchStep(
-            'lxc-hosts-ucf-non-interactive', **kwargs))
-    elif r.complete['osa-branch'] == 'stable/newton':
-        nextsteps.append(steps.PatchStep(
-            'lxc-hosts-ucf-non-interactive-newton', **kwargs))
-    else:
-        nextsteps.append(steps.PatchStep(
-            'lxc-hosts-ucf-non-interactive-ocata', **kwargs))
-
-    # Release specific steps: Mitaka
-    if r.complete['osa-branch'] == 'stable/mitaka' and utils.is_ironic(r):
-        nextsteps.append(
-            steps.FileAppendStep(
-                'enable-ironic',
-                '/etc/openstack_deploy/user_variables.yml',
-                '\n\nnova_virt_type: ironic\n',
-                **kwargs)
-            )
-
-    return nextsteps
-
-
-def stage8_ironic_networking(r, **kwargs):
-    """Configure all the special things for ironic networking."""
-
-    nextsteps = []
-
-    nextsteps.append(
-        steps.YamlAddElementStep(
-            'add-provider-network',
-            '/etc/openstack_deploy/openstack_user_config.yml',
-            ['global_overrides', 'provider_networks'],
-            {'network':
-                 {'group_binds': ['neutron_linuxbridge_agent',
-                                  'ironic_conductor_container',
-                                  'ironic_api_container'],
-                  'container_bridge': 'br-ironic',
-                  'container_type': 'veth',
-                  'container_interface': 'eth12',
-                  'type': 'flat',
-                  'net_name': 'ironic',
-                  'ip_from_q': 'ironic'
-                  }
-            },
-            **kwargs)
-        )
-
-    nextsteps.append(
-        steps.YamlDeleteElementStep(
-            'delete-provider-network',
-            '/etc/openstack_deploy/openstack_user_config.yml',
-            ['global_overrides', 'provider_networks'],
-            2,
-            **kwargs)
-        )
-
-    net, hosts = expand_ironic_netblock(r)
-
-    nextsteps.append(
-        steps.YamlUpdateElementStep(
-            'configure-external-lb-ip',
-            '/etc/openstack_deploy/openstack_user_config.yml',
-            ['global_overrides'],
-            'external_lb_vip_address',
-            str(hosts[4]),
-            **kwargs)
-        )
-
-    nextsteps.append(
-        steps.YamlUpdateDictionaryStep(
-            'add-network-cidr',
-            '/etc/openstack_deploy/openstack_user_config.yml',
-            ['cidr_networks'],
-            {'ironic': str(net)},
-            **kwargs)
-        )
-
-    nextsteps.append(
-        steps.YamlAddElementStep(
-            'reserve-netblock-start',
-            '/etc/openstack_deploy/openstack_user_config.yml',
-            ['used_ips'],
-            '%s,%s' % (hosts[0], hosts[10]),
-            **kwargs)
-        )
-    nextsteps.append(
-        steps.YamlAddElementStep(
-            'reserve-netblock-end',
-            '/etc/openstack_deploy/openstack_user_config.yml',
-            ['used_ips'],
-            '%s,%s' % (hosts[-10], hosts[-1]),
-            **kwargs)
-        )
-    nextsteps.append(
-        steps.SimpleCommandStep(
-            'add-ironic-bridge',
-            'brctl addbr br-ironic',
-            **kwargs)
-        )
-    nextsteps.append(
-        steps.SimpleCommandStep(
-            'add-ironic-bridge-nic',
-            'brctl addif br-ironic eth1',
-            **kwargs)
-        )
-    nextsteps.append(
-        steps.SimpleCommandStep(
-            'add-ironic-bridge-ip',
-            'ifconfig br-ironic inet %s up' % hosts[4],
-            **kwargs)
-        )
-    nextsteps.append(
-        steps.SimpleCommandStep(
-            'add-ironic-interface-ip',
-            'ifconfig eth1 inet %s up' % hosts[3],
-            **kwargs)
-        )
-    nextsteps.append(steps.PatchStep('ironic-vip-address', **r.kwargs))
-
-    return nextsteps
-
-
-def stage9_final_configuration(r, **kwargs):
-    """Final tweaks to configuration before we run the playbooks."""
-
-    nextsteps = []
-
-    # We also need to re-write git repos in a large number of roles
-    replacements = [
-        ('(http|https|git)://github.com',
-         r.complete['git-mirror-github']),
-        ('(http|https|git)://git.openstack.org',
-         r.complete['git-mirror-openstack']),
-        ('https://mirror.rackspace.com',
-         'http://mirror.rackspace.com'),
-        (' +checksum:.*', ''),
-        ]
-
-    if r.complete['local-cache'] != 'none':
-        replacements.append(
-            ('https://rpc-repo.rackspace.com',
-             'http://%s/rpc-repo.rackspace.com' % r.complete['local-cache'])
-            )
-        replacements.append(
-            ('https://bootstrap.pypa.io/get-pip.py',
-             'http://%s/pip/get-pip.py' % r.complete['local-cache'])
-            )
-
-    nextsteps.append(
-        steps.BulkRegexpEditorStep(
-            'bulk-edit-roles',
-            '/etc/ansible',
-            '.*\.(ini|yml)$',
-            replacements,
-            **kwargs))
-
-    # Release specific steps: Mitaka
-    if r.complete['osa-branch'] == 'stable/mitaka' and utils.is_ironic(r):
-        nextsteps.append(
-            steps.CopyFileStep(
-                'enable-ironic-environment-mitaka',
-                'etc/openstack_deploy/env.d/ironic.yml',
-                '/etc/openstack_deploy/env.d/ironic.yml',
-                **kwargs)
-            )
-
-    if utils.is_ironic(r):
-        nextsteps.append(steps.PatchStep('ironic-tftp-address', **kwargs))
-
-        if r.complete['osa-branch'] == 'stable/mitaka':
-            nextsteps.append(steps.PatchStep('ironic-pxe-options', **kwargs))
-        else:
-            nextsteps.append(steps.PatchStep('ironic-pxe-options-newton',
-                                             **kwargs))
-
-    return nextsteps
-
-
 def deploy(screen):
     global ARGS
 
@@ -403,7 +155,7 @@ def deploy(screen):
     r.resolve_steps(use_curses=(not ARGS.no_curses))
 
     if utils.is_ironic(r):
-        net, hosts = expand_ironic_netblock(r)
+        net, hosts = utils.expand_ironic_netblock(r)
         r.kwargs['max_attempts'] = 1
         r.load_step(steps.SimpleCommandStep(
                 'setup-neutron-ironic',
