+++ ostrich/utils.py
@@ -25,3 +25,12 @@ def recursive_dictionary_update(d, updates):
         else:
             d[key] = updates[key]
     return d
+
+
+def expand_ironic_netblock(r):
+    net = ipaddress.ip_network(r.complete['ironic-ip-block'])
+    hosts = []
+    for h in net.hosts():
+        hosts.append(str(h))
+
+    return net, hosts
